# perf: optimize `insert` and `insert_ser`