@jakmeier 2023-09-18
This is the follow up of https://github.com/near/nearcore/pull/9526